Move RemoveReference to settings view and change file name
