+++ client/src/js/references/components/Detail/Remove.js
