+++ client/src/js/references/components/Detail/__tests__/Remove.test.js
@@ -1,6 +1,6 @@
 import { RemoveBanner } from "../../../../base/index";
 import * as actions from "../../../actions";
-import RemoveReference from "../RemoveReference";
+import RemoveReference from "../Remove";
 
 describe("<RemoveReference />", () => {
     const initialState = {
